requirements.txt
transformers>=4.45.0
peft>=0.12.0
trl>=0.11.0
bitsandbytes>=0.43.0
accelerate>=0.34.0
datasets>=2.20.0
llama-cpp-python>=0.2.90
streamlit>=1.37.0
torch>=2.3.0
